+++ src/commands/index.ts
@@ -4,3 +4,4 @@ export { createLineCommand, createBarCommand } from "./lineBar";
 export { createBoxPlotCommand } from "./boxplot";
 export { createTableCommand } from "./table";
 export { createEntityBreakdownCommand } from "./entityBreakdown";
+export { createEntityMatrixCommand } from "./entityMatrix";